server/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
pymysql==1.1.1
pyodbc==5.1.0
gunicorn==23.0.0